moved input types to own file
