+++ app/modules/input.py
@@ -1,6 +1,6 @@
 import ctypes
 import time
-
+from app.types.input import Input, Input_I, KeyBdInput
 from app.utils.logger import log
 
 SendInput = ctypes.windll.user32.SendInput
@@ -9,46 +9,6 @@ SendInput = ctypes.windll.user32.SendInput
 KEYEVENTF_KEYDOWN = 0x0008
 KEYEVENTF_KEYUP = KEYEVENTF_KEYDOWN | 0x0002
 KEYEVENTF_EXTENDEDKEY = 0x0001
-
-# C struct redefinitions
-PUL = ctypes.POINTER(ctypes.c_ulong)
-
-
-class KeyBdInput(ctypes.Structure):
-    _fields_ = [
-        ("wVk", ctypes.c_ushort),
-        ("wScan", ctypes.c_ushort),
-        ("dwFlags", ctypes.c_ulong),
-        ("time", ctypes.c_ulong),
-        ("dwExtraInfo", PUL),
-    ]
-
-
-class HardwareInput(ctypes.Structure):
-    _fields_ = [
-        ("uMsg", ctypes.c_ulong),
-        ("wParamL", ctypes.c_short),
-        ("wParamH", ctypes.c_ushort),
-    ]
-
-
-class MouseInput(ctypes.Structure):
-    _fields_ = [
-        ("dx", ctypes.c_long),
-        ("dy", ctypes.c_long),
-        ("mouseData", ctypes.c_ulong),
-        ("dwFlags", ctypes.c_ulong),
-        ("time", ctypes.c_ulong),
-        ("dwExtraInfo", PUL),
-    ]
-
-
-class Input_I(ctypes.Union):
-    _fields_ = [("ki", KeyBdInput), ("mi", MouseInput), ("hi", HardwareInput)]
-
-
-class Input(ctypes.Structure):
-    _fields_ = [("type", ctypes.c_ulong), ("ii", Input_I)]
 
 
 class Key:
